+++ src/screens/Inside/ClientesList.jsx
@@ -2,6 +2,7 @@ import { StatusBar } from 'expo-status-bar';
 import { StyleSheet, Text, View } from 'react-native';
 
 export default function ClientesList() {
+  console.log('tela cliente list');
   return (
     <View style={styles.container}>
       <Text style={styles.title}>Cliente</Text>
