+++ src/screens/components/ListarComponents.jsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react';
 import { FlatList, TouchableOpacity, View, Text, Button, StyleSheet, SectionList } from 'react-native';
 import { addDoc, collection, query, getDocs, doc, updateDoc, deleteDoc, where } from 'firebase/firestore';
 import { useNavigation } from '@react-navigation/native';
-import { FontAwesome } from '@expo/vector-icons';
+import { FontAwesome, Feather } from '@expo/vector-icons';
 import { db } from '../../config/firebase';
 
 function Listar({ osList, selecionarOS }) {
@@ -11,39 +11,33 @@ function Listar({ osList, selecionarOS }) {
   const [osListState, setOSList] = useState([]);
   const osCollectionRef = collection(db, 'teste');
 
-  const [docId, setDocId] = useState('');
-
-  const navigateToOS = (os) => {
-    const osId = os.id;
-    navigation.push(`editOS/${osId}`); // Use the navigation prop to navigate to the editOS screen
-  };
-
-
-
-  useEffect(() => {
-    loadOS();
-  }, []);
+ 
 
   return (
     <SectionList
-      sections={[{ data: osListState }]}
-      renderItem={({ item }) => (
-        <TouchableOpacity onPress={() => navigateToOS(item)}> 
-          <View style={styles.container}>
-            <View style={styles.containerText}>
-              <Text>Status: {item.status}</Text>
-              <Text>Data: {item.data}</Text>
-              <Text>Cliente: {item.cliente}</Text>
-              <Text>Prioridade: {item.prioridade}</Text>
+      sections={[{ data: osList }]}
+      renderItem={({ item }) => {
+        console.log('Item:', item);
+  
+        return (
+          <TouchableOpacity onPress={() => navigateToOS(item)}>
+            <View style={styles.container}>
+              <View style={styles.containerText}>
+                <Text>Status: {item.status}</Text>
+                <Text>Data: {item.data}</Text>
+               
+                <Text>Cliente: {typeof item.cliente === 'object' ? item.cliente.value : item.cliente}</Text>
+              
+              </View>
+              <TouchableOpacity  onPress={() => navigateToOS(item)}
+                style={styles.button}>
+                <Feather name="edit" size={24} color="white" />
+              </TouchableOpacity>
+              
             </View>
-            <TouchableOpacity
-              onPress={() => navigateToOS(item)} 
-              style={styles.button}>
-              <FontAwesome name="pencil-square-o" size={24} color="black" />
-            </TouchableOpacity>
-          </View>
-        </TouchableOpacity>
-      )}
+          </TouchableOpacity>
+        );
+      }}
       keyExtractor={(item) => item.id.toString()}
     />
   );
@@ -67,6 +61,6 @@ const styles = StyleSheet.create({
     flexDirection: 'column',
   },
   button: {
-    marginLeft: 90,
+    marginLeft: 100,
   },
 });
